RAG_Project/ai-assistant/requirements.txt
streamlit
openai
langchain
langchain-openai
langchain-community
langchain-classic
faiss-cpu
pypdf
python-dotenv
ragas
langgraph
pandas
matplotlib
torch
transformers
peft
numexpr
tavily-python
